v2/backend/requirements.txt
# Web Framework
fastapi>=0.104.0  # Modern, fast web framework for building APIs
uvicorn[standard]>=0.24.0  # ASGI server for running FastAPI

# Database
sqlmodel>=0.0.14  # SQL database ORM built on SQLAlchemy and Pydantic
psycopg[binary]>=3.1.0  # PostgreSQL adapter for Python

# Configuration & Validation
pydantic>=2.5.0  # Data validation using Python type annotations
pydantic-settings>=2.1.0  # Settings management using Pydantic models
pydantic[email]>=2.5.0  # Email validation support (includes email-validator)
email-validator>=2.0.0  # Email validation for Pydantic EmailStr
python-dotenv>=1.0.0  # Load environment variables from .env files

# Authentication
python-jose[cryptography]>=3.3.0  # JWT token encoding/decoding
bcrypt>=4.0.0  # Password hashing with bcrypt (direct usage, not via passlib)

# File Upload
python-multipart>=0.0.6  # Support for multipart/form-data (file uploads)

# AWS S3 Storage
boto3>=1.29.0  # AWS SDK for Python (S3 access)

# AI/ML Services
replicate>=0.20.0  # Replicate API client for video generation and Whisper transcription

# Background Jobs
rq>=1.15.0  # Simple job queue for Python (Redis Queue)
redis>=5.0.0  # Redis client for Python

# Audio Processing
librosa>=0.10.0  # Audio analysis library (BPM, beats, tempo detection)
ffmpeg-python>=0.2.0  # Python bindings for FFmpeg (video/audio processing)

# Testing
pytest>=7.4.0  # Testing framework
pytest-asyncio>=0.21.0  # Pytest plugin for testing async code

